Module aptos_sdk::crypto::bls12381::bls12381_pop

This module provides APIs for proofs-of-possesion (PoPs) used to prevent rogue-key attacks, both for multisignatures and aggregate signatures.

Rogue-key attacks were first introduced by Micali, Ohta and Reyzin 1 and PoPs were first introduced by Ristenpart and Yilek 2.

Accountable-Subgroup Multisignatures: Extended Abstract; by Micali, Silvio and Ohta, Kazuo and Reyzin, Leonid; in Proceedings of the 8th ACM Conference on Computer and Communications Security; 2001; 1: The Power of Proofs-of-Possession: Securing Multiparty Signatures against Rogue-Key Attacks; by Ristenpart, Thomas and Yilek, Scott; in Advances in Cryptology - EUROCRYPT 2007; 2007 ↩

↩

Structs

ProofOfPossession

A proof-of-possesion (PoP) of a BLS12381 private key. This is just a BLS signature on the corresponding public key.

Constants

DST_BLS_POP_IN_G2

Domain separation tag (DST) for hashing a public key before computing its proof-of-possesion (PoP), which is also just a signature.