Trait aptos_sdk::crypto::hash::CryptoHasher

pub trait CryptoHasher: Default + Write {
    fn seed() -> &'static [u8; 32];
    fn update(&mut self, bytes: &[u8]);
    fn finish(self) -> HashValue;

    fn hash_all(bytes: &[u8]) -> HashValue { ... }
}

A trait for representing the state of a cryptographic hasher.

Required Methods

fn seed() -> &'static [u8; 32]

the seed used to initialize hashing Self before the serialization bytes of the actual value

fn update(&mut self, bytes: &[u8])

Write bytes into the hasher.

fn finish(self) -> HashValue

Finish constructing the HashValue.

Provided Methods

fn hash_all(bytes: &[u8]) -> HashValue

Convenience method to compute the hash of a complete byte slice.