Module aptos_sdk::crypto::multi_ed25519

This module provides an API for the accountable threshold multi-sig PureEdDSA signature scheme over the ed25519 twisted Edwards curve as defined in RFC8032.

Signature verification also checks and rejects non-canonical signatures.

Structs

MultiEd25519PrivateKey

Vector of private keys in the multi-key Ed25519 structure along with the threshold.

MultiEd25519PublicKey

Vector of public keys in the multi-key Ed25519 structure along with the threshold.

MultiEd25519Signature

Vector of the multi-key signatures along with a 32bit [u8; 4] bitmap required to map signatures with their corresponding public keys.

Constants

BITMAP_NUM_OF_BYTES

Number of bytes used for the bitmap in a MultiEd25519 private key

MAX_NUM_OF_KEYS

const for max number of ed25519 keys allowed for multi-ed25519 keys