Trait aptos_sdk::crypto::PublicKey

pub trait PublicKey: Clone + Eq + Hash + ValidCryptoMaterial + for<'a> From<&'a Self::PrivateKeyMaterial> {
    type PrivateKeyMaterial: PrivateKey
    where
        <Self::PrivateKeyMaterial as PrivateKey>::PublicKeyMaterial == Self;
}

A type for key material that can be publicly shared, and in asymmetric fashion, can be obtained from a PrivateKey reference. This convertibility requirement ensures the existence of a deterministic, canonical public key construction from a private key.

Required Associated Types

type PrivateKeyMaterial: PrivateKey
where
<Self::PrivateKeyMaterial as PrivateKey>::PublicKeyMaterial == Self

We require public / private types to be coupled, i.e. their associated type is each other.