Trait aptos_sdk::move_types::gas_algebra::ToUnitFractionalWithParams

pub trait ToUnitFractionalWithParams<U> {
    type Params;

    fn ratio(params: &Self::Params) -> (u64, u64);
}

Trait that defines a conversion from one unit to another, with a fractional conversion rate determined from the parameters dynamically.

Required Associated Types

type Params

Required Methods

fn ratio(params: &Self::Params) -> (u64, u64)