Module aptos_sdk::move_types::value

Structs

MoveFieldLayout

A layout associated with a named field

Enums

MoveStruct

MoveStructLayout

MoveTypeLayout

MoveValue

Constants

MOVE_STRUCT_FIELDS

In the WithTypes configuration, a Move struct gets serialized into a Serde struct with this as the second field

MOVE_STRUCT_NAME

In the WithTypes configuration, a Move struct gets serialized into a Serde struct with this name

MOVE_STRUCT_TYPE

In the WithTypes configuration, a Move struct gets serialized into a Serde struct with this as the first field

Functions

serialize_values