Module aptos_sdk::rest_client

Modules

aptos

aptos_api_types

error

faucet

response

state

types

Structs

Account

Client

FaucetClient

GasEstimationParams

IndexResponseBcs

The struct holding all data returned to the client by the index endpoint (i.e., GET “/”). This is just for the BCS response and cannot change

MoveModuleBytecode

Move module bytecode along with it’s ABI

PendingTransaction

A transaction waiting in mempool

Resource

Response

State

VersionedNewBlockEvent

Enums

Transaction

Enum of the different types of transactions in Aptos

Constants

DEFAULT_VERSION_PATH_BASE

USER_AGENT

Functions

deserialize_from_prefixed_hex_string

retriable

retriable_with_404