Function aptos_sdk::transaction_builder::aptos_stdlib::account_offer_signer_capability

pub fn account_offer_signer_capability(
    signer_capability_sig_bytes: Vec<u8, Global>, 
    account_scheme: u8, 
    account_public_key_bytes: Vec<u8, Global>, 
    recipient_address: AccountAddress
) -> TransactionPayload

Offers signer capability on behalf of account to the account at address recipient_address. An account can delegate its signer capability to only one other address at one time. signer_capability_key_bytes is the SignerCapabilityOfferProofChallengeV2 signed by the account owner’s key account_scheme is the scheme of the account (ed25519 or multi_ed25519) account_public_key_bytes is the public key of the account owner recipient_address is the address of the recipient of the signer capability - note that if there’s an existing recipient_address in the account owner’s SignerCapabilityOffer, this will replace the previous recipient_address upon successful verification (the previous recipient will no longer have access to the account owner’s signer capability)