Module aptos_sdk::types

Modules

access_path

Suppose we have the following data structure in a smart contract:

account_address

account_config

account_state

account_view

block_info

block_metadata

chain_id

contract_event

epoch_change

epoch_state

event

ledger_info

mempool_status

move_resource

network_address

nibble

Nibble represents a four-bit unsigned integer.

on_chain_config

proof

serde_helper

state_proof

state_store

timestamp

transaction

trusted_state

utility_coin

validator_config

validator_info

validator_signer

validator_verifier

vm_status

waypoint

write_set

For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.

Structs

AccountKey

LocalAccount

PeerId

A struct that represents an account address.

Statics

TEST_COIN_TYPE