Module aptos_sdk::types::account_config

Modules

constants

events

resources

Structs

AccountResource

A Rust representation of an Account resource. This is not how the Account is represented in the VM but it’s a convenient representation.

Aggregator

Rust representation of Aggregator Move struct.

BlockResource

Should be kept in-sync with BlockResource move struct in block.move.

ChainIdResource

CoinInfoResource

Rust representation of CoinInfo Move resource.

CoinStoreResource

The balance resource held under an account.

DepositEvent

Struct that represents a DepositPaymentEvent.

Integer

Rust representation of Integer Move struct.

NewBlockEvent

Struct that represents a NewBlockEvent. Should be kept in-sync with NewBlockEvent move struct in block.move.

NewEpochEvent

Struct that represents a NewEpochEvent.

OptionalAggregator

Rust representation of OptionalAggregator Move struct.

TransactionValidation

A Rust representation of chain-specific account information

WithdrawEvent

Struct that represents a SentPaymentEvent.

Enums

AccountSequenceInfo

Constants

APTOS_ACCOUNT_MODULE_IDENTIFIER

The Identifier for the Account module.

CORE_ACCOUNT_MODULE_IDENTIFIER

The Identifier for the Account module.

CORE_CODE_ADDRESS

Hex address: 0x1

EVENT_MODULE_IDENTIFIER

Statics

APTOS_ACCOUNT_MODULE

The ModuleId for the Account module.

APTOS_TRANSACTION_VALIDATION

CORE_ACCOUNT_MODULE

The ModuleId for the Account module.

EVENT_MODULE

NEW_BLOCK_EVENT_PATH

The path to the new block event handle under a Block::BlockResource resource.

Functions

aptos_test_root_address

new_block_event_key

reserved_vm_address