Trait aptos_sdk::types::move_resource::MoveStorage

pub trait MoveStorage {
    fn fetch_resource_by_version(
        &self, 
        access_path: AccessPath, 
        version: u64
    ) -> Result<Vec<u8, Global>, Error>;
    fn fetch_config_by_version(
        &self, 
        config_id: ConfigID, 
        version: u64
    ) -> Result<Vec<u8, Global>, Error>;
    fn fetch_synced_version(&self) -> Result<u64, Error>;
    fn fetch_latest_state_checkpoint_version(&self) -> Result<u64, Error>;
}

Required Methods

fn fetch_resource_by_version(
&self,
access_path: AccessPath,
version: u64
) -> Result<Vec<u8, Global>, Error>

Returns a Move resources as serialized byte array from a specified version of the database.

fn fetch_config_by_version(
&self,
config_id: ConfigID,
version: u64
) -> Result<Vec<u8, Global>, Error>

Returns an on-chain resource as a serialized byte array from a specified version of the database.

fn fetch_synced_version(&self) -> Result<u64, Error>

Get the version on the latest transaction info.

fn fetch_latest_state_checkpoint_version(&self) -> Result<u64, Error>

Get the version of the latest state checkpoint