Trait aptos_sdk::types::on_chain_config::OnChainConfig

pub trait OnChainConfig: Send + Sync + DeserializeOwned {
    const MODULE_IDENTIFIER: &'static str;
    const TYPE_IDENTIFIER: &'static str;
    const ADDRESS: &'static str = ;
    const CONFIG_ID: ConfigID = const CONFIG_ID: ConfigID =
    ConfigID(Self::ADDRESS, Self::MODULE_IDENTIFIER, Self::TYPE_IDENTIFIER);;

    fn deserialize_default_impl(bytes: &[u8]) -> Result<Self, Error> { ... }
    fn deserialize_into_config(bytes: &[u8]) -> Result<Self, Error> { ... }
    fn fetch_config<T>(storage: &T) -> Option<Self>
    where
        T: ConfigStorage,
    { ... }
    fn access_path() -> AccessPath { ... }
    fn struct_tag() -> StructTag { ... }
}

Trait to be implemented by a Rust struct representation of an on-chain config that is stored in storage as a serialized byte array

Required Associated Constants

const MODULE_IDENTIFIER: &'static str

const TYPE_IDENTIFIER: &'static str

Provided Associated Constants

const ADDRESS: &'static str =

const CONFIG_ID: ConfigID = const CONFIG_ID: ConfigID = ConfigID(Self::ADDRESS, Self::MODULE_IDENTIFIER, Self::TYPE_IDENTIFIER);

Provided Methods

fn deserialize_default_impl(bytes: &[u8]) -> Result<Self, Error>

fn deserialize_into_config(bytes: &[u8]) -> Result<Self, Error>

fn fetch_config<T>(storage: &T) -> Option<Self> where
T: ConfigStorage,

fn access_path() -> AccessPath

fn struct_tag() -> StructTag