Module aptos_sdk::types::proof::definition

This module has definition of various proofs.

Structs

AccumulatorConsistencyProof

A proof that can be used to show that two Merkle accumulators are consistent – the big one can be obtained by appending certain leaves to the small one. For example, at some point in time a client knows that the root hash of the ledger at version 10 is old_root (it could be a waypoint). If a server wants to prove that the new ledger at version N is derived from the old ledger the client knows, it can show the subtrees that represent all the new leaves. If the client can verify that it can indeed obtain the new root hash by appending these new leaves, it can be convinced that the two accumulators are consistent.

AccumulatorExtensionProof

A proof that first verifies that establishes correct computation of the root and then returns the new tree to acquire a new root and version.

AccumulatorProof

A proof that can be used authenticate an element in an accumulator given trusted root hash. For example, both LedgerInfoToTransactionInfoProof and TransactionInfoToEventProof can be constructed on top of this structure.

AccumulatorRangeProof

A proof that is similar to AccumulatorProof, but can be used to authenticate a range of leaves. For example, given the following accumulator:

SparseMerkleProof

A proof that can be used to authenticate an element in a Sparse Merkle Tree given trusted root hash. For example, TransactionInfoToAccountProof can be constructed on top of this structure.

SparseMerkleProofExt

A more detailed version of SparseMerkleProof with the only difference that all the leaf siblings are explicitly set as SparseMerkleLeafNode instead of its hash value.

SparseMerkleRangeProof

Note: this is not a range proof in the sense that a range of nodes is verified! Instead, it verifies the entire left part of the tree up to a known rightmost node. See the description below.

TransactionAccumulatorSummary

An in-memory accumulator for storing a summary of the core transaction info accumulator. It is a summary in the sense that it only stores maximally frozen subtree nodes rather than storing all leaves and internal nodes.

TransactionInfoListWithProof

The proof used to authenticate a list of consecutive transaction infos.

TransactionInfoWithProof

TransactionInfo and a TransactionAccumulatorProof connecting it to the ledger root.

Enums

NodeInProof

Constants

MAX_ACCUMULATOR_LEAVES

MAX_ACCUMULATOR_PROOF_DEPTH

Type Definitions

EventAccumulatorProof

LeafCount

Because leaves can only take half the space in the tree, any numbering of the tree leaves must not take the full width of the total space. Thus, for a 64-bit ordering, our maximumm proof depth is limited to 63.

TransactionAccumulatorProof

TransactionAccumulatorRangeProof