Struct aptos_sdk::types::proof::position::FrozenSubTreeIterator

pub struct FrozenSubTreeIterator { /* private fields */ }

Traverse leaves from left to right in groups that forms full subtrees, yielding root positions of such subtrees. Note that each 1-bit in num_leaves corresponds to a full subtree. For example, in the below tree of 5=0b101 leaves, the two 1-bits corresponds to Fzn2 and L4 accordingly.

           Non-fzn
          /       \
         /         \
        /           \
      Fzn2         Non-fzn
     /   \           /   \
    /     \         /     \
   Fzn1    Fzn3  Non-fzn  [Placeholder]
  /  \    /  \    /    \
 L0  L1  L2  L3 L4   [Placeholder]

Implementations

impl FrozenSubTreeIterator

pub fn new(num_leaves: u64) -> FrozenSubTreeIterator