Module aptos_sdk::types

Modules

access_path

Suppose we have the following data structure in a smart contract:

account_address

account_config

account_state

account_view

aggregate_signature

block_info

block_metadata

chain_id

contract_event

epoch_change

epoch_state

event

governance

ledger_info

mempool_status

move_resource

network_address

nibble

Nibble represents a four-bit unsigned integer.

on_chain_config

proof

serde_helper

stake_pool

staking_contract

state_proof

state_store

timestamp

transaction

trusted_state

utility_coin

validator_config

validator_info

validator_performances

validator_signer

validator_verifier

vesting

vm_status

waypoint

write_set

For each transaction the VM executes, the VM will output a WriteSet that contains each access path it updates. For each access path, the VM can either give its new value or delete it.

Structs

AccountKey

LocalAccount

LocalAccount represents an account on the Aptos blockchain. Internally it holds the private / public key pair and the address of the account. You can use this struct to help transact with the blockchain, e.g. by generating a new account and signing transactions.

PeerId

A struct that represents an account address.

Statics

APTOS_COIN_TYPE